Move repo updater scheduler into separate subpackage
This PR decreases the bloat of the internal/repos package slightly by moving the update scheduler into a separate package.
