+++ cmd/repo-updater/internal/repoupdater/BUILD.bazel
@@ -60,10 +60,12 @@ go_test(
         "//internal/extsvc/awscodecommit",
         "//internal/extsvc/github",
         "//internal/extsvc/gitlab",
+        "//internal/gitserver",
         "//internal/grpc",
         "//internal/grpc/defaults",
         "//internal/observation",
         "//internal/repos",
+        "//internal/repos/scheduler",
         "//internal/repoupdater",
         "//internal/repoupdater/protocol",
         "//internal/repoupdater/v1:repoupdater",
