+++ cmd/repo-updater/shared/BUILD.bazel
@@ -31,6 +31,7 @@ go_library(
         "//internal/encryption/keyring",
         "//internal/env",
         "//internal/extsvc",
+        "//internal/gitserver",
         "//internal/goroutine",
         "//internal/goroutine/recorder",
         "//internal/grpc",
@@ -41,6 +42,7 @@ go_library(
         "//internal/observation",
         "//internal/ratelimit",
         "//internal/repos",
+        "//internal/repos/scheduler",
         "//internal/repoupdater/v1:repoupdater",
         "//internal/service",
         "//internal/timeutil",
