+++ internal/repos/BUILD.bazel
@@ -33,7 +33,6 @@ go_library(
         "python_packages.go",
         "ruby_packages.go",
         "rust_packages.go",
-        "scheduler.go",
         "sources.go",
         "sources_test_utils.go",
         "status_messages.go",
@@ -81,19 +80,16 @@ go_library(
         "//internal/extsvc/pypi",
         "//internal/extsvc/rubygems",
         "//internal/gitserver",
-        "//internal/gitserver/protocol",
         "//internal/goroutine",
         "//internal/httpcli",
         "//internal/httptestutil",
         "//internal/jsonc",
         "//internal/lazyregexp",
         "//internal/licensing",
-        "//internal/limiter",
         "//internal/metrics",
         "//internal/observation",
         "//internal/ratelimit",
         "//internal/repoupdater",
-        "//internal/repoupdater/protocol",
         "//internal/trace",
         "//internal/types",
         "//internal/types/typestest",
@@ -149,7 +145,6 @@ go_test(
         "pagure_test.go",
         "perforce_test.go",
         "python_packages_test.go",
-        "scheduler_test.go",
         "sources_test.go",
         "status_messages_test.go",
         "store_test.go",
@@ -189,12 +184,10 @@ go_test(
         "//internal/extsvc/phabricator",
         "//internal/github_apps/types",
         "//internal/gitserver",
-        "//internal/gitserver/protocol",
         "//internal/goroutine",
         "//internal/httpcli",
         "//internal/httptestutil",
         "//internal/licensing",
-        "//internal/limiter",
         "//internal/observation",
         "//internal/ratelimit",
         "//internal/rcache",
@@ -205,7 +198,6 @@ go_test(
         "//lib/errors",
         "//lib/pointers",
         "//schema",
-        "@com_github_davecgh_go_spew//spew",
         "@com_github_dnaeon_go_vcr//cassette",
         "@com_github_dnaeon_go_vcr//recorder",
         "@com_github_google_go_cmp//cmp",
