+++ internal/repos/syncer_test.go
@@ -171,7 +171,7 @@ func TestSyncerSync(t *testing.T) {
 		svcs    []*types.ExternalService
 		ctx     context.Context
 		now     func() time.Time
-		diff    repos.Diff
+		diff    types.RepoSyncDiff
 		err     string
 	}
 
@@ -197,7 +197,7 @@ func TestSyncerSync(t *testing.T) {
 				store:  store,
 				stored: types.Repos{},
 				now:    clock.Now,
-				diff: repos.Diff{Added: types.Repos{tc.repo.With(
+				diff: types.RepoSyncDiff{Added: types.Repos{tc.repo.With(
 					typestest.Opt.RepoCreatedAt(clock.Time(1)),
 					typestest.Opt.RepoSources(tc.svc.Clone().URN()),
 				)}},
@@ -206,7 +206,7 @@ func TestSyncerSync(t *testing.T) {
 			},
 		)
 
-		var diff repos.Diff
+		var diff types.RepoSyncDiff
 		diff.Unmodified = append(diff.Unmodified, tc.repo.With(
 			typestest.Opt.RepoSources(tc.svc.URN()),
 		))
@@ -241,7 +241,7 @@ func TestSyncerSync(t *testing.T) {
 					typestest.Opt.RepoSources(tc.svc.URN()),
 				)},
 				now: clock.Now,
-				diff: repos.Diff{
+				diff: types.RepoSyncDiff{
 					Deleted: types.Repos{
 						tc.repo.With(func(r *types.Repo) {
 							r.Sources = map[string]*types.SourceInfo{}
@@ -282,7 +282,7 @@ func TestSyncerSync(t *testing.T) {
 					typestest.Opt.RepoSources(tc.svc.URN()),
 				)},
 				now: clock.Now,
-				diff: repos.Diff{
+				diff: types.RepoSyncDiff{
 					Deleted: types.Repos{
 						tc.repo.With(func(r *types.Repo) {
 							r.Sources = map[string]*types.SourceInfo{}
@@ -337,7 +337,7 @@ func TestSyncerSync(t *testing.T) {
 					typestest.Opt.RepoSources(tc.svc.URN()),
 				)},
 				now: clock.Now,
-				diff: repos.Diff{Unmodified: types.Repos{tc.repo.With(
+				diff: types.RepoSyncDiff{Unmodified: types.Repos{tc.repo.With(
 					typestest.Opt.RepoSources(tc.svc.URN()),
 				)}},
 				svcs: []*types.ExternalService{tc.svc},
@@ -355,7 +355,7 @@ func TestSyncerSync(t *testing.T) {
 					typestest.Opt.RepoSources(tc.svc.URN()),
 				)},
 				now: clock.Now,
-				diff: repos.Diff{Unmodified: types.Repos{tc.repo.With(
+				diff: types.RepoSyncDiff{Unmodified: types.Repos{tc.repo.With(
 					typestest.Opt.RepoSources(tc.svc.URN()),
 				)}},
 				svcs: []*types.ExternalService{tc.svc},
@@ -373,7 +373,7 @@ func TestSyncerSync(t *testing.T) {
 					typestest.Opt.RepoSources(tc.svc.URN(), svcdup.URN()),
 				)},
 				now: clock.Now,
-				diff: repos.Diff{Unmodified: types.Repos{tc.repo.With(
+				diff: types.RepoSyncDiff{Unmodified: types.Repos{tc.repo.With(
 					typestest.Opt.RepoSources(tc.svc.URN(), svcdup.URN()),
 				)}},
 				svcs: []*types.ExternalService{tc.svc},
@@ -389,7 +389,7 @@ func TestSyncerSync(t *testing.T) {
 					typestest.Opt.RepoSources(tc.svc.URN(), svcdup.URN()),
 				)},
 				now: clock.Now,
-				diff: repos.Diff{Deleted: types.Repos{tc.repo.With(
+				diff: types.RepoSyncDiff{Deleted: types.Repos{tc.repo.With(
 					typestest.Opt.RepoDeletedAt(clock.Time(1)),
 				)}},
 				svcs: []*types.ExternalService{tc.svc, &svcdup},
@@ -403,8 +403,8 @@ func TestSyncerSync(t *testing.T) {
 					r.Name = "old-name"
 				})},
 				now: clock.Now,
-				diff: repos.Diff{
-					Modified: repos.ReposModified{
+				diff: types.RepoSyncDiff{
+					Modified: types.ReposModified{
 						{
 							Repo:     tc.repo.With(typestest.Opt.RepoModifiedAt(clock.Time(1))),
 							Modified: types.RepoModifiedName,
@@ -430,7 +430,7 @@ func TestSyncerSync(t *testing.T) {
 					}),
 				},
 				now: clock.Now,
-				diff: repos.Diff{
+				diff: types.RepoSyncDiff{
 					Deleted: types.Repos{
 						tc.repo.With(func(r *types.Repo) {
 							r.Sources = map[string]*types.SourceInfo{}
@@ -438,7 +438,7 @@ func TestSyncerSync(t *testing.T) {
 							r.UpdatedAt = clock.Time(0)
 						}),
 					},
-					Modified: repos.ReposModified{
+					Modified: types.ReposModified{
 						{
 							Repo: tc.repo.With(
 								typestest.Opt.RepoModifiedAt(clock.Time(1)),
@@ -463,7 +463,7 @@ func TestSyncerSync(t *testing.T) {
 					tc.repo.With(typestest.Opt.RepoExternalID("another-id")),
 				},
 				now: clock.Now,
-				diff: repos.Diff{
+				diff: types.RepoSyncDiff{
 					Added: types.Repos{
 						tc.repo.With(
 							typestest.Opt.RepoCreatedAt(clock.Time(1)),
@@ -495,8 +495,8 @@ func TestSyncerSync(t *testing.T) {
 					tc.repo.With(typestest.Opt.RepoExternalID("bar")), // same name as sourced
 				}.With(typestest.Opt.RepoCreatedAt(clock.Time(1))),
 				now: clock.Now,
-				diff: repos.Diff{
-					Modified: repos.ReposModified{
+				diff: types.RepoSyncDiff{
+					Modified: types.ReposModified{
 						{
 							Repo: tc.repo.With(
 								typestest.Opt.RepoCreatedAt(clock.Time(1)),
@@ -528,7 +528,7 @@ func TestSyncerSync(t *testing.T) {
 					r.DeletedAt = clock.Time(0)
 				})},
 				now: clock.Now,
-				diff: repos.Diff{Added: types.Repos{
+				diff: types.RepoSyncDiff{Added: types.Repos{
 					tc.repo.With(
 						typestest.Opt.RepoCreatedAt(clock.Time(1))),
 				}},
@@ -559,8 +559,8 @@ func TestSyncerSync(t *testing.T) {
 						r.ExternalRepo.ID = "2"
 					}),
 				},
-				diff: repos.Diff{
-					Modified: repos.ReposModified{
+				diff: types.RepoSyncDiff{
+					Modified: types.ReposModified{
 						{
 							Repo: tc.repo.With(func(r *types.Repo) {
 								r.Name = "foo"
@@ -591,8 +591,8 @@ func TestSyncerSync(t *testing.T) {
 					tc.repo.With(typestest.Opt.RepoName(api.RepoName(strings.ToUpper(string(tc.repo.Name))))),
 				},
 				now: clock.Now,
-				diff: repos.Diff{
-					Modified: repos.ReposModified{
+				diff: types.RepoSyncDiff{
+					Modified: types.ReposModified{
 						{Repo: tc.repo.With(typestest.Opt.RepoModifiedAt(clock.Time(0)))},
 					},
 				},
@@ -726,12 +726,12 @@ func TestSyncRepo(t *testing.T) {
 	testCases := []struct {
 		name       string
 		repo       api.RepoName
-		background bool        // whether to run SyncRepo in the background
-		before     types.Repos // the repos to insert into the database before syncing
-		sourced    *types.Repo // the repo that is returned by the fake sourcer
-		returned   *types.Repo // the expected return value from SyncRepo (which changes meaning depending on background)
-		after      types.Repos // the expected database repos after syncing
-		diff       repos.Diff  // the expected repos.Diff sent by the syncer
+		background bool               // whether to run SyncRepo in the background
+		before     types.Repos        // the repos to insert into the database before syncing
+		sourced    *types.Repo        // the repo that is returned by the fake sourcer
+		returned   *types.Repo        // the expected return value from SyncRepo (which changes meaning depending on background)
+		after      types.Repos        // the expected database repos after syncing
+		diff       types.RepoSyncDiff // the expected types.Diff sent by the syncer
 	}{{
 		name:       "insert",
 		repo:       repo.Name,
@@ -739,7 +739,7 @@ func TestSyncRepo(t *testing.T) {
 		sourced:    repo.Clone(),
 		returned:   repo,
 		after:      types.Repos{repo},
-		diff: repos.Diff{
+		diff: types.RepoSyncDiff{
 			Added: types.Repos{repo},
 		},
 	}, {
@@ -750,8 +750,8 @@ func TestSyncRepo(t *testing.T) {
 		sourced:    repo.Clone(),
 		returned:   oldRepo,
 		after:      types.Repos{repo},
-		diff: repos.Diff{
-			Modified: repos.ReposModified{
+		diff: types.RepoSyncDiff{
+			Modified: types.ReposModified{
 				{Repo: repo, Modified: types.RepoModifiedStars},
 			},
 		},
@@ -763,8 +763,8 @@ func TestSyncRepo(t *testing.T) {
 		sourced:    repo.Clone(),
 		returned:   repo,
 		after:      types.Repos{repo},
-		diff: repos.Diff{
-			Modified: repos.ReposModified{
+		diff: types.RepoSyncDiff{
+			Modified: types.ReposModified{
 				{Repo: repo, Modified: types.RepoModifiedStars},
 			},
 		},
@@ -776,8 +776,8 @@ func TestSyncRepo(t *testing.T) {
 		sourced:    repo.Clone(),
 		returned:   repo,
 		after:      types.Repos{repo},
-		diff: repos.Diff{
-			Modified: repos.ReposModified{
+		diff: types.RepoSyncDiff{
+			Modified: types.ReposModified{
 				{Repo: repo, Modified: types.RepoModifiedName},
 			},
 		},
@@ -789,8 +789,8 @@ func TestSyncRepo(t *testing.T) {
 		sourced:    repo.With(typestest.Opt.RepoArchived(true)),
 		returned:   repo,
 		after:      types.Repos{repo.With(typestest.Opt.RepoArchived(true))},
-		diff: repos.Diff{
-			Modified: repos.ReposModified{
+		diff: types.RepoSyncDiff{
+			Modified: types.ReposModified{
 				{
 					Repo:     repo.With(typestest.Opt.RepoArchived(true)),
 					Modified: types.RepoModifiedArchived,
@@ -805,8 +805,8 @@ func TestSyncRepo(t *testing.T) {
 		sourced:    repo.Clone(),
 		returned:   repo.With(typestest.Opt.RepoArchived(true)),
 		after:      types.Repos{repo},
-		diff: repos.Diff{
-			Modified: repos.ReposModified{
+		diff: types.RepoSyncDiff{
+			Modified: types.ReposModified{
 				{Repo: repo, Modified: types.RepoModifiedArchived},
 			},
 		},
@@ -818,8 +818,8 @@ func TestSyncRepo(t *testing.T) {
 		sourced:    repo.Clone(),
 		returned:   repo.With(typestest.Opt.RepoExternalID("old id")),
 		after:      types.Repos{repo},
-		diff: repos.Diff{
-			Modified: repos.ReposModified{
+		diff: types.RepoSyncDiff{
+			Modified: types.ReposModified{
 				{Repo: repo, Modified: types.RepoModifiedExternalRepo},
 			},
 		},
@@ -834,8 +834,8 @@ func TestSyncRepo(t *testing.T) {
 		sourced:  repo.Clone(),
 		returned: repo.With(typestest.Opt.RepoExternalID("old id")),
 		after:    types.Repos{repo},
-		diff: repos.Diff{
-			Modified: repos.ReposModified{
+		diff: types.RepoSyncDiff{
+			Modified: types.ReposModified{
 				{Repo: repo, Modified: types.RepoModifiedName},
 			},
 		},
@@ -862,7 +862,7 @@ func TestSyncRepo(t *testing.T) {
 				ObsvCtx: observation.TestContextTB(t),
 				Now:     time.Now,
 				Store:   store,
-				Synced:  make(chan repos.Diff, 1),
+				Synced:  make(chan types.RepoSyncDiff, 1),
 				Sourcer: repos.NewFakeSourcer(nil,
 					repos.NewFakeSource(servicesPerKind[extsvc.KindGitHub], nil, tc.sourced),
 				),
@@ -943,7 +943,7 @@ func TestSyncRun(t *testing.T) {
 		ObsvCtx: observation.TestContextTB(t),
 		Sourcer: repos.NewFakeSourcer(nil, fakeSource),
 		Store:   store,
-		Synced:  make(chan repos.Diff),
+		Synced:  make(chan types.RepoSyncDiff),
 		Now:     time.Now,
 	}
 
@@ -969,7 +969,7 @@ func TestSyncRun(t *testing.T) {
 	// The first thing sent down Synced is the list of repos in store during
 	// initialisation
 	diff := <-syncer.Synced
-	if d := cmp.Diff(repos.Diff{Unmodified: stored}, diff, ignore); d != "" {
+	if d := cmp.Diff(types.RepoSyncDiff{Unmodified: stored}, diff, ignore); d != "" {
 		t.Fatalf("Synced mismatch (-want +got):\n%s", d)
 	}
 
@@ -981,8 +981,8 @@ func TestSyncRun(t *testing.T) {
 
 	// Next up it should find the existing repo and send it down Synced
 	diff = <-syncer.Synced
-	if d := cmp.Diff(repos.Diff{
-		Modified: repos.ReposModified{
+	if d := cmp.Diff(types.RepoSyncDiff{
+		Modified: types.ReposModified{
 			{Repo: sourced[0], Modified: types.RepoModifiedDescription},
 		},
 	}, diff, ignore); d != "" {
@@ -991,7 +991,7 @@ func TestSyncRun(t *testing.T) {
 
 	// Then the new repo.
 	diff = <-syncer.Synced
-	if d := cmp.Diff(repos.Diff{Added: sourced[1:]}, diff, ignore); d != "" {
+	if d := cmp.Diff(types.RepoSyncDiff{Added: sourced[1:]}, diff, ignore); d != "" {
 		t.Fatalf("Synced mismatch (-want +got):\n%s", d)
 	}
 
@@ -1002,12 +1002,12 @@ func TestSyncRun(t *testing.T) {
 	// We check synced again to test us going around the Run loop 2 times in
 	// total.
 	diff = <-syncer.Synced
-	if d := cmp.Diff(repos.Diff{Unmodified: sourced[:1]}, diff, ignore); d != "" {
+	if d := cmp.Diff(types.RepoSyncDiff{Unmodified: sourced[:1]}, diff, ignore); d != "" {
 		t.Fatalf("Synced mismatch (-want +got):\n%s", d)
 	}
 
 	diff = <-syncer.Synced
-	if d := cmp.Diff(repos.Diff{Unmodified: sourced[1:]}, diff, ignore); d != "" {
+	if d := cmp.Diff(types.RepoSyncDiff{Unmodified: sourced[1:]}, diff, ignore); d != "" {
 		t.Fatalf("Synced mismatch (-want +got):\n%s", d)
 	}
 
@@ -1103,7 +1103,7 @@ func TestSyncerMultipleServices(t *testing.T) {
 			return s, nil
 		},
 		Store:  store,
-		Synced: make(chan repos.Diff),
+		Synced: make(chan types.RepoSyncDiff),
 		Now:    time.Now,
 	}
 
@@ -1123,7 +1123,7 @@ func TestSyncerMultipleServices(t *testing.T) {
 
 	// The first thing sent down Synced is an empty list of repos in store.
 	diff := <-syncer.Synced
-	if d := cmp.Diff(repos.Diff{}, diff, ignore); d != "" {
+	if d := cmp.Diff(types.RepoSyncDiff{}, diff, ignore); d != "" {
 		t.Fatalf("initial Synced mismatch (-want +got):\n%s", d)
 	}
 
@@ -1521,7 +1521,7 @@ func TestConflictingSyncers(t *testing.T) {
 			return s, nil
 		},
 		Store:  tx2,
-		Synced: make(chan repos.Diff, 2),
+		Synced: make(chan types.RepoSyncDiff, 2),
 		Now:    time.Now,
 	}
 
@@ -1930,7 +1930,7 @@ func TestSyncReposWithLastErrors(t *testing.T) {
 			diff := <-syncer.Synced
 
 			deleted := types.Repos{&types.Repo{ID: dbRepos[0].ID}}
-			if d := cmp.Diff(repos.Diff{Deleted: deleted}, diff); d != "" {
+			if d := cmp.Diff(types.RepoSyncDiff{Deleted: deleted}, diff); d != "" {
 				t.Fatalf("Deleted mismatch (-want +got):\n%s", d)
 			}
 
@@ -2034,7 +2034,7 @@ func setupSyncErroredTest(ctx context.Context, s repos.Store, t *testing.T,
 		ObsvCtx: observation.TestContextTB(t),
 		Now:     time.Now,
 		Store:   s,
-		Synced:  make(chan repos.Diff, 1),
+		Synced:  make(chan types.RepoSyncDiff, 1),
 		Sourcer: repos.NewFakeSourcer(
 			nil,
 			repos.NewFakeSource(&service,
